requirements.txt
#generals
Django>=5.0,<6.0
mysqlclient>=2.2.0
psycopg[binary]>=3.1.19
python-dotenv>=1.0.0
django-bootstrap5>=23.0
djangorestframework

#third parties
markdown    
django-filter
drf-yasg[validation]